requirements.txt
spacy>=3.4.0
boto3>=1.24.0
python-dotenv>=0.19.0 